Require `NTH_ELSE` rather than `else` following expectations.
It was unfortunately easy to write `NTH_EXPECT(...) { ... }` (forgetting
the `else`). This looks like it does the same thing but actually
evaluates the expectation and then unconditionally evaluates the
compound statement. As far as I can tell, there's no workaround for this
problem other than using a macro in place of the `else` to hide code
that won't compile if it's forgotten.

Even so, the error message Clang gives isn't great, and kinda points you
in the wrong direction. We'll have to experiment with other thnigs that
can be hidden in there.